requirements.txt
psycopg2
django-cors-headers
pyyaml
django
dj-database-url
django-heroku
djangorestframework
markdown
django-filter
gunicorn
whitenoise
django_extensions
graphviz
dotplus
pytest
mypy
pycodestyle